rewrite class components into pure functions
